requirements.txt
altair==5.0.1
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
cryptography==41.0.1
dateparser==1.1.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.1
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
parso==0.8.3
pdf2image==1.16.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
PyPDF2==3.0.1
pypng==0.20220715.0
pyrsistent==0.19.3
pytesseract==0.3.10
python-barcode==0.14.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
qrcode==7.4.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rich==13.4.2
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.23.1
tenacity==8.2.2
tesseract==0.1.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.15.0
